Refactoring the codebase and updating `leaflet` and `react-reaflet`
