+++ src/Controls/Controls.js
@@ -1,15 +1,13 @@
+// noinspection NpmUsedModulesInstalled,JSUnresolvedVariable
+
 import MainControls from './MainControls';
 import BehaviorControls from './BehaviorControls';
 
-// noinspection JSUnresolvedVariable
 import {__} from '@wordpress/i18n';
-// noinspection JSUnresolvedVariable
 const {PanelBody} = wp.components;
-// noinspection JSUnresolvedVariable
 const {InspectorControls} = wp.blockEditor;
 
 export default function Controls({props}) {
-	// noinspection JSXNamespaceValidation
 	return (
 		<InspectorControls>
 			<PanelBody
