+++ src/deprecated.js
@@ -70,7 +70,6 @@ const deprecated = [
 					scrollWheelZoom,
 				},
 			} = props;
-			// noinspection JSXNamespaceValidation
 			return markers ? (
 				<div className={className}>
 					<div className="ootb-openstreetmap--map"
