+++ src/edit.js
@@ -2,6 +2,7 @@ import Controls from './Controls/Controls';
 import SearchBox from './Elements/SearchBox';
 import LeafletMap from './Elements/LeafletMap';
 import Alert from './Elements/Alert';
+// noinspection NpmUsedModulesInstalled
 import {useBlockProps} from '@wordpress/block-editor';
 
 export default function edit(props) {
